Fix the refresh rate reported for the Desktop video mode.
This was broken in r25457. I've also taken the opportunity to move the
rate variable to a float for increased accuracy.


git-svn-id: http://svn.mythtv.org/svn/trunk@25983 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/mythplayer.cpp
@@ -1596,7 +1596,7 @@ void MythPlayer::InitAVSync(void)
 
     repeat_delay = 0;
 
-    refreshrate = MythDisplay::GetDisplayInfo().rate;
+    refreshrate = (int)MythDisplay::GetDisplayInfo().rate;
     if (refreshrate <= 0)
         refreshrate = frame_interval;
 
@@ -1994,7 +1994,7 @@ void MythPlayer::VideoStart(void)
 
     float temp_speed = (play_speed == 0.0) ? audio.GetStretchFactor() : play_speed;
     uint fr_int = (int)(1000000.0 / video_frame_rate / temp_speed);
-    uint rf_int = MythDisplay::GetDisplayInfo().rate;
+    uint rf_int = (int)MythDisplay::GetDisplayInfo().rate;
 
     // Default to Interlaced playback to allocate the deinterlacer structures
     // Enable autodetection of interlaced/progressive from video stream
